fix: use 'object' as default classification (#138)
The REST API operations that return collections of objects include a
classification property. Many nodes in the tree come back with a value
of "" for this property. However those nodes should be classified with
"object" instead. This is being resolved in the REST API; however this
commit addresses the issue when this client is used with older
unpatched Metasys devices.